worker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bayoo-docx==0.2.20
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
exceptiongroup==1.2.0
fastapi==0.110.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.6
latex2mathml==3.77.0
lxml==5.1.0
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
OOoPy==2.0
pandas==2.0.3
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive==1.3.1
PyDrive2==1.19.0
pygsheets==2.0.6
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yamllint==1.35.1
celery
redis